Znosi gorny limit kolejki enrichmentu.

+++ backend/database/migrations/2026_07_20_110000_add_enrichment_batch_limit_to_ai_settings.php
@@ -11,7 +11,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('ai_settings', function (Blueprint $table): void {
-            $table->unsignedSmallInteger('enrichment_batch_limit')
+            $table->unsignedInteger('enrichment_batch_limit')
                 ->default(5)
                 ->after('tavily_search_mode');
         });
